Changed units_convert to int

+++ src/rappture_io.f90
@@ -1199,7 +1199,7 @@ subroutine readTriggeredFzMode(  jumpDeflLim, PiezoReverseTime_pct , Want_FzEig,
   call read_integer(driver, INPUT_PREFIX // "(op).choice(fzshape).current" , fzshape , InputEcho, "fzshape")
   PiezoReverseTime_pct = readGenericDbl(status, INPUT_PREFIX // "(op).number(PiezoReverseTime_pct).current", "PiezoReverseTime_pct", InputEcho, 1)
 
-  jumpDeflLim = readGenericDbl(status, INPUT_PREFIX // "(op).number(jumpDeflLim).current", "jumpDeflLim", InputEcho, 1.0)
+  jumpDeflLim = readGenericDbl(status, INPUT_PREFIX // "(op).number(jumpDeflLim).current", "jumpDeflLim", InputEcho, 1)
   jumpDeflLim = jumpDeflLim / 1e9;
 
   Want_FzEig = daniel_get_boolean( INPUT_PREFIX // "(sim).group(plots).boolean(Want_FzEig).current")
@@ -3003,7 +3003,7 @@ real*8 function readGenericDbl( status, element_to_read, short_name, InputEcho, 
   integer, intent(inout) :: status
   character(len=*), intent(in) :: element_to_read, short_name
   character*1500, intent(inout) :: InputEcho
-  real*8, intent(in) :: units_convert
+  integer, intent(in) :: units_convert
 
   character*100 :: strVal
   integer :: rp_units_convert_dbl
